requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.36.0
torch==2.1.1
python-dotenv==1.0.0
accelerate==0.25.0
pydantic==2.5.0
numpy==1.24.3
huggingface-hub==0.19.4
requests==2.31.0